requirements.txt
dis-snek @ git+https://github.com/Discord-Snake-Pit/Dis-Snek.git
pymysql
python-dotenv